+++ include/pickle.h
@@ -27,7 +27,7 @@ public:
         this->nvlink_handles_.push_back(handle);
     }
 
-    [[nodiscard]] std::shared_ptr<Event> send(uint32_t unique_id, uint64_t addr, uint32_t length);
+    [[nodiscard]] std::shared_ptr<Event> send(uint32_t unique_id, uint64_t addr, uint64_t length);
 
     void poll();
 
@@ -39,8 +39,8 @@ private:
     explicit PickleSender(std::shared_ptr<RdmaSender> rdma_sender);
     explicit PickleSender(std::shared_ptr<NvlinkSender> nvlink_sender);
 
-    uint32_t lookup_lkey(uint64_t addr, uint32_t length) const;
-    NvlinkHandle lookup_ipc_handle(uint64_t addr, uint32_t length) const;
+    uint32_t lookup_lkey(uint64_t addr, uint64_t length) const;
+    NvlinkHandle lookup_ipc_handle(uint64_t addr, uint64_t length) const;
 
     BackendType backend_type_;
     std::shared_ptr<RdmaSender> rdma_sender_;
@@ -60,7 +60,7 @@ public:
     // 注册一个用于 RDMA rkey 查找的 MemoryRegion
     void register_memory_region(std::shared_ptr<rdma_util::MemoryRegion> mr);
 
-    [[nodiscard]] std::shared_ptr<Event> recv(uint32_t unique_id, uint64_t addr, uint32_t length);
+    [[nodiscard]] std::shared_ptr<Event> recv(uint32_t unique_id, uint64_t addr, uint64_t length);
 
     void poll();
 
@@ -72,7 +72,7 @@ private:
     explicit PickleRecver(std::shared_ptr<RdmaRecver> rdma_recver);
     explicit PickleRecver(std::shared_ptr<NvlinkRecver> nvlink_recver);
 
-    uint32_t lookup_rkey(uint64_t addr, uint32_t length) const;
+    uint32_t lookup_rkey(uint64_t addr, uint64_t length) const;
 
     BackendType backend_type_;
     std::shared_ptr<RdmaRecver> rdma_recver_;
