+++ tests/test_nvlink_executor.cpp
@@ -64,7 +64,7 @@ int run_child_recver() {
 
     uint32_t uid = 1234;
 
-    auto ev_recv = recver->recv(uid, reinterpret_cast<uint64_t>(dst), static_cast<uint32_t>(N * sizeof(int)));
+    auto ev_recv = recver->recv(uid, reinterpret_cast<uint64_t>(dst), static_cast<uint64_t>(N * sizeof(int)));
 
     // 5. 后台轮询匹配 + 触发 copy 任务
     while (!ev_recv->is_notified()) {
@@ -142,7 +142,7 @@ int run_parent_sender(pid_t child_pid) {
     // 5. 发起 NVLink 发送
     // The sender API expects an offset from the IPC handle base pointer; the buffer
     // starts at the base, so offset should be 0 instead of passing the raw pointer.
-    auto ev_send = sender->send(uid, /*offset=*/0, static_cast<uint32_t>(N * sizeof(int)), handle);
+    auto ev_send = sender->send(uid, /*offset=*/0, static_cast<uint64_t>(N * sizeof(int)), handle);
 
     // 6. 轮询 ACK，直到发送侧事件完成
     while (!ev_send->is_notified()) {
